Here is my final project readme and images
Final Project Assignment
Created by
Andrew Whitmire
First off, I have been having many issues trying to get everything to work. I implemented a Image in the app. I know everything worked so I followed the other coding, so it should work. I could not get any emulator to work. I tried uninstalling the software and reinstalling from scratch as a fresh install on the whole program. I tried updating everything and nothing seemed to work. I tried adding a database and nothing would work because the compiler in build.grade (module file) is no longer relevant at the end of 2018. It kept giving me an error. I could not do everything I wanted because of the issues and not knowing if it works. Below is an image for the coding. It starts with the activity_main.xml and the only other file which is java.